Katamari: finished the core game, but still todo - unlink "game" size from the collision size, adjust the speeds, adjust the sizes

+++ L4-Katamari/Davork/InputDevice.cpp
@@ -67,8 +67,8 @@ void InputDevice::OnMouseMove(RawMouseEventArgs args)
 	GetCursorPos(&p);
 	ScreenToClient(game->Display->hWnd, &p);
 	
-	MousePosition	= Vector2(p.x, p.y);
-	MouseOffset		= Vector2(args.X, args.Y);
+	MousePosition	= Vector2(static_cast<float>(p.x), static_cast<float>(p.y));
+	MouseOffset		= Vector2(static_cast<float>(args.X), static_cast<float>(args.Y));
 	MouseWheelDelta = args.WheelDelta;
 
 	const MouseMoveEventArgs moveArgs = {MousePosition, MouseOffset, MouseWheelDelta};
